Figured out how to split the graph service properly
Also fixed a few issues found after the split. Still need to do more testing, and fix ownership changing when importing.
